+++ frontend/src/global.tsx
@@ -1,5 +1,4 @@
 import {Button, message, notification} from 'antd';
-import {useIntl} from 'umi';
 import defaultSettings from '../config/defaultSettings';
 
 const {pwa} = defaultSettings;
@@ -9,7 +8,7 @@ const isHttps = document.location.protocol === 'https:';
 if (pwa) {
   // Notify user if offline now
   window.addEventListener('sw.offline', () => {
-    message.warning(useIntl().formatMessage({id: 'app.pwa.offline'}));
+    message.warning('当前处于离线状态');
   });
 
   // Pop up a prompt on the page asking the user if they want to use the latest version
@@ -35,7 +34,7 @@ if (pwa) {
         worker.postMessage({type: 'skip-waiting'}, [channel.port2]);
       });
       // Refresh current page to use the updated HTML and other assets after SW has skiped waiting
-      window.location.reload(true);
+      window.location.reload();
       return true;
     };
     const key = `open${Date.now()}`;
@@ -47,12 +46,12 @@ if (pwa) {
           reloadSW();
         }}
       >
-        {useIntl().formatMessage({id: 'app.pwa.serviceworker.updated.ok'})}
+        刷新
       </Button>
     );
     notification.open({
-      message: useIntl().formatMessage({id: 'app.pwa.serviceworker.updated'}),
-      description: useIntl().formatMessage({id: 'app.pwa.serviceworker.updated.hint'}),
+      message: '有新内容',
+      description: '请点击「刷新」按钮或手动刷新页面',
       btn,
       key,
       onClose: async () => null,